refactor mm to cm

+++ src/main/java/dat/backend/model/entities/ExtraDiameterMaterial.java
@@ -3,7 +3,7 @@ package dat.backend.model.entities;
 public abstract class ExtraDiameterMaterial extends AMaterial {
     private int diameter;
 
-    public ExtraDiameterMaterial(int id, String name, int length, double price, String unit, int diameter) {
+    public ExtraDiameterMaterial(int id, String name, double length, double price, String unit, int diameter) {
         super(id, name, length, price, unit);
         this.diameter = diameter;
     }
