+++ src/main/java/dat/backend/model/entities/Fitting.java
@@ -2,7 +2,7 @@ package dat.backend.model.entities;
 
 public class Fitting extends ExtraDimensionMaterial {
 
-    public Fitting(int id, String name, int length, double price, String unit, int width, int height) {
+    public Fitting(int id, String name, double length, double price, String unit, double width, double height) {
         super(id, name, length, price, unit, width, height);
     }
 }
