+++ src/main/java/dat/backend/model/entities/RoofTile.java
@@ -2,7 +2,7 @@ package dat.backend.model.entities;
 
 public class RoofTile extends ExtraDimensionMaterial {
 
-        public RoofTile(int id, String name, int length, double price, String unit, int width) {
+        public RoofTile(int id, String name, double length, double price, String unit, double width) {
             super(id, name, length, price, unit, width);
         }
