+++ src/main/java/dat/backend/model/entities/Screw.java
@@ -3,7 +3,7 @@ package dat.backend.model.entities;
 public class Screw extends ExtraDiameterMaterial{
 
 
-        public Screw(int id, String name, int length, double price, String unit, int diameter) {
+        public Screw(int id, String name, double length, double price, String unit, int diameter) {
             super(id, name, length, price, unit, diameter);
         }
 }
