+++ src/main/java/dat/backend/model/entities/Wood.java
@@ -4,7 +4,7 @@ public class Wood extends ExtraDimensionMaterial{
     private String category;
     private boolean isPressureTreated;
 
-    public Wood(int id, String name, int length, double price, String unit, String category, double width, double height, boolean isPressureTreated) {
+    public Wood(int id, String name, double length, double price, String unit, String category, double width, double height, boolean isPressureTreated) {
         super(id, name, length, price, unit, width, height);
         this.category = category;
         this.isPressureTreated = isPressureTreated;
